AddNewStuff
Add new code for adding miles to your auto's and sql for creating the DB
and tables for the program.

+++ Models/Gas.cs
@@ -9,7 +9,7 @@ namespace mvc_MilesTracker.Models
     {
         public string AutoName = string.Empty;
         public int Miles = 0;
-        public float Gallons = 0;
-        public double Price = 0.0;
+        public double Gallons = 0;
+        public decimal Price = 0;
     }
 }
